Basic positioning of seshview
Basic css positioning of the seshview is done.
The number of comments is purposefully commented out.
The acutal date ie Aug 30 is currently only 30,
I have not been able to get the month in name.
Might want to use moment.js which is bundled
with Kalendae.
There seems to be a bug with the number of users